ftdi_eeprom: Fix git build problems by using autoreconf
For recent versions of automake, the autotools steps in Makefile.cvs
don't work:
  $ make -f Makefile.cvs
  aclocal
  autoheader
  automake
  configure.in: required file `./install-sh' not found
  configure.in: required file `./mkinstalldirs' not found
  configure.in: required file `./missing' not found
  src/Makefile.am: required file `./depcomp' not found
  make: *** [all] Error 1

This is because automake no longer calls libtoolize, as described in:
  http://www.gnu.org/software/automake/manual/html_node/Error-required-file-ltmain_002esh-not-found.html
The recommended workaround there is to use autoreconf rather than
calling individual autotools programs.

This patch also removes the use of all_includes and all_libraries, as
they are undefined and cause automake errors.

Signed-off-by: Jim Paris <jim@jtan.com>